feat(BA-7289): move the v2-consumed specs to models/specs
Declare every spec type the v2 actions and ops consume next to the
schema: DataQuerier, DataLookup (renamed from DataFinder along with the
lookup_data operation), Searcher/SearcherResult, the pagination
strategies, and the pure-ABC DataUpdater/DataBatchUpdater lineage.
repositories/base keeps legacy-compatible bridge subclasses for the
transition and drops its now-empty pagination and searcher modules; spec
consumers pick up the typed to_data(row) signature.
