+++ src/ai/backend/manager/models/specs/types.py
@@ -1,4 +1,5 @@
-"""Check declarations write specs carry, consumed by the ops execution layer."""
+"""Shared declarations of the v2 lineage: the checks write specs carry, and the
+result shapes the ops execution answers with."""
 
 from __future__ import annotations
 
@@ -6,6 +7,7 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
 from ai.backend.common.exception import BackendAIError
+from ai.backend.common.identifier.entity import EntityID
 from ai.backend.manager.models.clauses import QueryCondition
 
 if TYPE_CHECKING:
@@ -44,3 +46,25 @@ class IntegrityErrorCheck:
 
     constraint_name: str | None = None
     """Optional constraint name filter. If None, matches any constraint of the given type."""
+
+
+@dataclass
+class BulkResultWithFailures[TData]:
+    """What a bulk write did to each entity the caller named.
+
+    Named as the atomic bulk results are, minus a spec name it cannot carry — one type
+    serves both the updater and the purger — leaving the part a caller has to know:
+    some of these may have failed while the rest went through.
+
+    Keyed by entity rather than positional: the bulk shape answers per entity, and an
+    answer attached to the wrong one is worse than no answer.
+
+    Ordering is per group, not the caller's. A caller that needs its own order re-reads
+    these by the ids it passed in.
+
+    Names its fields as the other bulk results do, so a caller reading ``successes`` and
+    ``errors`` off this reads them the same way off ``BulkUpdaterResult``.
+    """
+
+    successes: dict[EntityID, TData]
+    errors: dict[EntityID, Exception]
