+++ src/ai/backend/manager/repositories/base/querier.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
 from collections.abc import Sequence
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any, Protocol, TypeVar
@@ -14,8 +13,7 @@ from ai.backend.manager.errors.repository import UnsupportedCompositePrimaryKeyE
 from ai.backend.manager.models.base import Base
 from ai.backend.manager.models.clauses import QueryCondition, QueryOrder
 from ai.backend.manager.models.scopes import ExistenceCheck, OperationScope
-
-from .pagination import PageInfoResult, QueryPagination
+from ai.backend.manager.models.specs.pagination import PageInfoResult, QueryPagination
 
 if TYPE_CHECKING:
     from sqlalchemy.engine import Row
@@ -47,99 +45,6 @@ class QuerierResult[TRow: Base]:
     """Result of executing a single-row query operation."""
 
     row: TRow
-
-
-class DataFinder[TRow: Base, TData](ABC):
-    """Reads one entity by a key that is not its primary key.
-
-    A lookup resolves an external key — a name, an email within a domain, a canonical
-    plus an architecture — and those are unique constraints rather than primary keys, so
-    :class:`Querier` cannot express them: it carries a single ``pk_value`` and derives
-    the WHERE from the table's primary key.
-
-    Conditions rather than a SELECT, unlike :class:`~...searcher.Searcher`, so one spec
-    stays one table. A key that needs a join — an image resolved through its alias table
-    — is a domain repository method, not this.
-
-    What separates it from a search is the expected cardinality, which is why
-    ``find_data`` reads at most two rows and rejects the second: matching more than one
-    means the key is not unique, and answering with an arbitrary one would hide that.
-
-    Example:
-        class UserByEmail(DataFinder[UserRow, UserData]):
-            def row_class(self) -> type[UserRow]:
-                return UserRow
-
-            def conditions(self) -> Sequence[QueryCondition]:
-                return [
-                    lambda: UserRow.email == self._email,
-                    lambda: UserRow.domain_name == self._domain,
-                ]
-
-            def to_data(self, row: UserRow) -> UserData:
-                return row.to_data()
-
-        async with ops.read_ops() as r:
-            user = await r.find_data(UserByEmail(email, domain))
-    """
-
-    @abstractmethod
-    def row_class(self) -> type[TRow]:
-        """Return the ORM class the key resolves within."""
-        raise NotImplementedError
-
-    @abstractmethod
-    def conditions(self) -> Sequence[QueryCondition]:
-        """Return the key's conditions, AND combined."""
-        raise NotImplementedError
-
-    @abstractmethod
-    def to_data(self, row: TRow) -> TData:
-        """Convert the matched row into its ``data/`` type."""
-        raise NotImplementedError
-
-
-class DataQuerier[TRow: Base, TData](ABC):
-    """Reads one entity: which row to fetch, and how that row becomes data.
-
-    Self-contained counterpart of :class:`Querier`, in the same sense
-    :class:`~ai.backend.manager.repositories.base.searcher.Searcher` is one for batch
-    queries. Carrying ``to_data`` is the point: the ops layer returns the ``data/`` type
-    and the ORM row never leaves it, so no caller has to know that rows have a
-    ``to_data`` at all.
-
-    Subclasses live in the domain repository, which is what lets ``to_data`` name its
-    row class directly.
-
-    Example:
-        class UserQuerier(DataQuerier[UserRow, UserData]):
-            def row_class(self) -> type[UserRow]:
-                return UserRow
-
-            def pk_value(self) -> UUID:
-                return self._user_id
-
-            def to_data(self, row: UserRow) -> UserData:
-                return row.to_data()
-
-        async with ops.read_ops() as r:
-            user = await r.query_data(UserQuerier(user_id))
-    """
-
-    @abstractmethod
-    def row_class(self) -> type[TRow]:
-        """Return the ORM class for table access and PK detection."""
-        raise NotImplementedError
-
-    @abstractmethod
-    def pk_value(self) -> UUID | str | int:
-        """Return the primary key value identifying the target row."""
-        raise NotImplementedError
-
-    @abstractmethod
-    def to_data(self, row: TRow) -> TData:
-        """Convert the fetched row into its ``data/`` type."""
-        raise NotImplementedError
 
 
 async def execute_querier[TRow: Base](
