+++ src/ai/backend/manager/repositories/base/types.py
@@ -3,11 +3,12 @@
 from __future__ import annotations
 
 from collections.abc import Callable
-from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any, TypeVar
 
-from ai.backend.common.identifier.entity import EntityID
 from ai.backend.manager.models.clauses import QueryCondition
+from ai.backend.manager.models.specs.types import (
+    BulkResultWithFailures as BulkResultWithFailures,
+)
 
 # Moved to models/specs/types.py with the v2 spec lineage; re-imported here for the
 # legacy spec modules that still import them from this path.
@@ -21,25 +22,3 @@ if TYPE_CHECKING:
 type CursorConditionFactory = Callable[[str], QueryCondition]
 
 TRow = TypeVar("TRow", bound="Row[Any]")
-
-
-@dataclass
-class BulkResultWithFailures[TData]:
-    """What a bulk write did to each entity the caller named.
-
-    Named as the atomic bulk results are, minus a spec name it cannot carry — one type
-    serves both the updater and the purger — leaving the part a caller has to know:
-    some of these may have failed while the rest went through.
-
-    Keyed by entity rather than positional: the bulk shape answers per entity, and an
-    answer attached to the wrong one is worse than no answer.
-
-    Ordering is per group, not the caller's. A caller that needs its own order re-reads
-    these by the ids it passed in.
-
-    Names its fields as the other bulk results do, so a caller reading ``successes`` and
-    ``errors`` off this reads them the same way off ``BulkUpdaterResult``.
-    """
-
-    successes: dict[EntityID, TData]
-    errors: dict[EntityID, Exception]
